requirements.txt
flask
werkzeug
openai
python-docx
pdfplumber
beautifulsoup4
requests